# api.yaml
openapi: 3.0.1
info:
  title: defaultTitle
  description: defaultDescription
  version: '0.1'
servers:
  - url: http://localhost:8080
paths:
  /registerNewBook:
    post:
      description: Auto generated using Swagger Inspector
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: number
                stock:
                  type: integer
                bookName:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "bookName":"Alamut Kalesi",
                      "stock":1,
                      "price":15.00
                  }
      responses:
        '201':
          description: Auto generated using Swagger Inspector
          content:
            text/plain;charset=UTF-8:
              schema:
                type: string
              examples: { }
      servers:
        - url: http://localhost:8080
    servers:
      - url: http://localhost:8080
...
  /updateBookStock:
    post:
      description: Auto generated using Swagger Inspector
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                price:
                  type: number
                stock:
                  type: integer
                bookName:
                  type: string
            examples:
              '0':
                value: |-
                  {
                      "bookName":"Alamut Kalesi",
                      "stock":1,
                      "price":15.00
                  }
      responses:
        '200':
          description: Auto generated using Swagger Inspector
          content:
            text/plain;charset=UTF-8:
              schema:
                type: string
              examples: { }
      servers:
        - url: http://localhost:8080
    servers:
      - url: http://localhost:8080
...